[REFACTOR] #101 - TabBar 컴포넌트로 분리
